python/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ragtab"
version = "0.1.0"
description = "Trích xuất bảng từ ảnh/PDF thành Markdown cho RAG pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
keywords = ["table-extraction", "ocr", "unet", "rag"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "torch>=2.0",
    "torchvision",
    "numpy",
    "opencv-python",
    "Pillow",
    "paddleocr",          
    "scipy",
    "matplotlib",        
    "timm",
]

[project.optional-dependencies]
dev = ["pytest", "build", "twine"]

[project.urls]
Homepage = "https://github.com/tai03102004/rag-table"
Repository = "https://github.com/tai03102004/rag-table"

[tool.setuptools.packages.find]
where = ["python"]
include = ["ragtab*"]